requirements.txt
markdown2
python-liquid
python-frontmatter
pyyaml